frontend/requirements.txt
pandas
plotly
python-dotenv
requests
streamlit